Move qec_round_iterator to util
